+++ src/ru/JavaKanban/utils/Managers.java
@@ -1,0 +1,5 @@
+package ru.JavaKanban.utils;
+
+public class Managers {
+  
+}
